+++ world/ObjectItem.go
@@ -39,7 +39,7 @@ func (o *ObjectItem) setArchetype(targetArch *data.Archetype) {
 	o.name, _ = targetArch.Name.GetString()*/
 }
 
-func (o *ObjectItem) update(d int) {
+func (o *ObjectItem) update(d int64) {
 }
 
 func (o *ObjectItem) getType() cdata.ArchetypeType {
